+ codegen executionpayload/transaction, - handwritten
